chore: move OpeningBook into eve/match
